Fixed dependency issue
The ExternalModuleDependencies field in the manifest is for external
dependencies from the perspective of the PSGallery or similar
repository. When this is specified it essentially skips the automatic
install of dependencies and instead warns the user if their local
system does not have the required module.

Also added a GUID to the dependency to ensure potential conflicts with
other repository sources.