Fixed UI windowing logic
If the number of items in $SelectedItemMembersToShow is
greater than the default size of 4, the windowing logic would
behave abnormally. The logic has been update to account for
how many members were actually specified.

NOTE: The vertical spacing could be further optimized to remove
the blank space at the bottom of the UI.