Added SelectionFormat parameter
This parameter offers different formatting for the final value
assigned to $Selection.

Note: There is no handling for Ctrl+C events for this logic. This means
the resulting $Selections output may not be in the expected format
when the user presses Ctrl+C (more than likely it will still be in the
default Boolean[] format as that is the form it uses to process the
UI events).