Changed handling of selections variable
The user provided selections reference variable will now be
set to null upon entering the function and only assigned the
the final value at the conclusion of the operation.

This may be able to be improved on further to fully support
Ctrl+C events. But this is a reasonable middle ground for now.