Added visual cue for column rendering
If not all columns can be rendered within the UI width, the right-bar
of the UI will be replaced with double vertical bars. While this
indication may be a bit too subtle, this is a simple solution to help
a user determine if there is any benefit to increasing the width of
the UI (without wasting additional space that could otherwise be
used for displaying content).